@@ -1,21 +1,21 @@
 {
   "name": "Example",
-  "version": "0.0.2",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "test": "jest"
   },
   "dependencies": {
-    "react": "16.0.0-alpha.12",
-    "react-native": "0.48.1",
-    "react-native-android-wifi": "0.0.19"
+    "react": "16.2.0",
+    "react-native": "0.53.0",
+    "react-native-android-wifi": "0.0.23"
   },
   "devDependencies": {
-    "babel-jest": "21.0.0",
-    "babel-preset-react-native": "3.0.2",
-    "jest": "21.0.1",
-    "react-test-renderer": "16.0.0-alpha.12"
+    "babel-jest": "22.2.0",
+    "babel-preset-react-native": "4.0.0",
+    "jest": "22.2.0",
+    "react-test-renderer": "16.2.0"
   },
   "jest": {
     "preset": "react-native"
